+++ changelog.md
@@ -4,6 +4,7 @@ vX.X.X
 - Added filter field in TX/RX sub-tab, closes #84.
 - Fixed incorrect link to GitHub on homepage, closes #94.
 - Fixed incorrect web page title on homepage, closes #93.
+- TX and RX buffer sizes are now shown on stats sub-tab, closes #95.
 
 v0.4.1
 ------